Fix build of nxkbd on an AMD64 host computer. On an AMD64, sizeof(long)==8. This code expected sizeof(long)==4.

+++ user/microwin/src/mwin/bmp/convbmp.c
@@ -29,8 +29,8 @@ typedef unsigned char	BYTE;
 
 #if !_MINIX
 typedef unsigned short	WORD;
-typedef unsigned long	DWORD;
-typedef long		LONG;
+typedef unsigned int	DWORD;
+typedef int		LONG;
 #define	CASTWORD
 #define	CASTDWORD
 #define	CASTLONG
